feat(core): clicking tray icon on Windows will bring windows to front
Doesn't make sense to do this on mac or linux but useful for Windows.
